requirements.txt
# Pinned dependency versions — generated from development environment
# Update by running: pip freeze | grep -iE '^(package)==' > requirements.txt
SQLAlchemy==2.0.45
chromadb==1.5.6
networkx==3.4.2
numpy==2.3.5
PyYAML==6.0.3
openai==2.30.0
alembic==1.18.4

# Dev dependencies
pytest==9.0.2
pytest-asyncio==1.3.0
